Reorganize build files
