+++ src/generators/generate-clone-reducer.js
@@ -17,21 +17,7 @@
 'use strict';
 
 const spec = require('shift-spec').default;
-
-function isStatefulType(type) {
-  switch (type.typeName) {
-    case 'Enum':
-    case 'String':
-    case 'Number':
-    case 'Boolean':
-      return false;
-    case 'Maybe':
-    case 'List':
-      return isStatefulType(type.argument);
-    default:
-      return true;
-  }
-}
+const { parameterize, isStatefulType } = require('../utilities.js');
 
 let content = `/**
  * Copyright 2016 Shape Security, Inc.
@@ -49,51 +35,37 @@ let content = `/**
  * limitations under the License.
  */
 
-const director = {`;
+import * as Shift from 'shift-ast';
 
-function reduce(name, type) {
-  switch (type.typeName) {
-    case 'Maybe':
-      return `${name} && ${reduce(name, type.argument)}`;
-    case 'List':
-      return `${name}.map(v => ${reduce('v', type.argument)})`;
-    case 'Union':
-      return `this[${name}.type](reducer, ${name})`;
-    default:
-      return `this.${type.typeName}(reducer, ${name})`;
+export default class CloneReducer {`;
+
+function cloneField(f) {
+  if (!isStatefulType(f.type)) {
+    return `${f.name}: node.${f.name}`;
   }
+  return parameterize(f.name);
 }
 
 for (let [typeName, type] of Object.entries(spec)) {
-  let fields = type.fields.filter(f => f.name !== 'type' && isStatefulType(f.type));
+  let fields = type.fields.filter(f => f.name !== 'type');
   if (fields.length === 0) {
     content += `
-  ${typeName}(reducer, node) {
-    return reducer.reduce${typeName}(node);
-  },
+  reduce${typeName}(node) {
+    return new Shift.${typeName};
+  }
 `;
-
   } else {
+    let statefulFields = fields.filter(f => isStatefulType(f.type));
+    let param = statefulFields.length > 0 ? `, { ${statefulFields.map(f => parameterize(f.name)).join(', ')} }` : '';
     content += `
-  ${typeName}(reducer, node) {
-    return reducer.reduce${typeName}(node, { ${fields.map(f => `${f.name}: ${reduce(`node.${f.name}`, f.type)}`).join(', ')} });
-  },
+  reduce${typeName}(node${param}) {
+    return new Shift.${typeName}({ ${fields.map(cloneField).join(', ')} });
+  }
 `;
-
   }
 }
 
-content += `};
-
-
-export default function reduce(reducer, node) {
-  return director[node.type](reducer, node);
-}
-
-export { default as CloneReducer } from './clone-reducer';
-export { default as LazyCloneReducer } from './lazy-clone-reducer';
-export { default as MonoidalReducer } from './monoidal-reducer';
-
+content += `}
 `;
 
-require('fs').writeFile('gen/index.js', content, 'utf-8');
+require('fs').writeFile('gen/clone-reducer.js', content, 'utf-8');
