+++ src/generators/generate-lazy-clone-reducer.js
@@ -16,46 +16,8 @@
 
 'use strict';
 
-let spec = require('shift-spec').default;
-const { isRestrictedWord, isReservedWordES6 } = require('esutils').keyword;
-
-function parameterize(fieldName) {
-  if (isRestrictedWord(fieldName) || isReservedWordES6(fieldName)) {
-    return '_' + fieldName;
-  }
-  return fieldName;
-}
-
-function asParameter(fieldName) {
-  if (isRestrictedWord(fieldName) || isReservedWordES6(fieldName)) {
-    return fieldName + ': _' + fieldName;
-  }
-  return fieldName;
-}
-
-function isNodeOrUnionOfNodes(type) {
-  return type.typeName === 'Union' && type.arguments.every(isNodeOrUnionOfNodes) || spec.hasOwnProperty(type.typeName);
-}
-
-function isStatefulType(type) {
-  switch (type.typeName) {
-    case 'Enum':
-    case 'String':
-    case 'Number':
-    case 'Boolean':
-      return false;
-    case 'Maybe':
-    case 'List':
-      return isStatefulType(type.argument);
-    case 'Union':
-      return type.arguments.some(isStatefulType);
-    default:
-      if (isNodeOrUnionOfNodes(type)) {
-        return true;
-      }
-      throw new Error('unimplemented: type ' + type);
-  }
-}
+const spec = require('shift-spec').default;
+const { sanitize, parameterize, isStatefulType, isNodeOrUnionOfNodes } = require('../utilities.js');
 
 function generateEquals(type, a, b) {
   switch (type.typeName) {
@@ -121,13 +83,13 @@ function cloneField(f) {
   if (!isStatefulType(f.type)) {
     return `${f.name}: node.${f.name}`;
   }
-  return asParameter(f.name);
+  return parameterize(f.name);
 }
 
 for (let [typeName, type] of Object.entries(spec)) {
   let fields = type.fields.filter(f => f.name !== 'type');
   let statefulFields = fields.filter(f => isStatefulType(f.type));
-  let param = statefulFields.length > 0 ? `, { ${statefulFields.map(f => asParameter(f.name)).join(', ')} }` : '';
+  let param = statefulFields.length > 0 ? `, { ${statefulFields.map(f => parameterize(f.name)).join(', ')} }` : '';
   content += `
   reduce${typeName}(node${param}) {`;
   if (statefulFields.length === 0) {
@@ -137,7 +99,7 @@ for (let [typeName, type] of Object.entries(spec)) {
 `;
   } else {
     content += `
-    if (${statefulFields.map(f => generateEquals(f.type, 'node.' + f.name, parameterize(f.name))).join(' && ')}) {
+    if (${statefulFields.map(f => generateEquals(f.type, 'node.' + f.name, sanitize(f.name))).join(' && ')}) {
       return node;
     }
     return new Shift.${typeName}({ ${fields.map(cloneField).join(', ')} });
